workshop2/news-gcs-agent-service/requirements.txt
# news_gcs_agent/requirements.txt
google-adk
python-dotenv
uvicorn
fastapi
nltk




